Fix compiler warning

+++ MZFayeClient/MZFayeClient.m
@@ -319,7 +319,7 @@ NSInteger const MZFayeClientDefaultMaximumAttempts = 5;
 {
     self.sentMessageCount++;
 
-    return [[NSString stringWithFormat:@"%d",self.sentMessageCount] base64String];
+    return [[NSString stringWithFormat:@"%ld", (long)self.sentMessageCount] base64String];
 }
 
 #pragma mark - Public methods
